refactor(folder): restructure flight modules into controllers, services, utils, and helpers
